UDPSetUp
This commit contains a chat log file, with no GUI. A udp client and
server which can send messages via udp back and forth to each other. The
serverthread is just a beginning of making it multithreaded/multiclient.
Usernames are implement by the client class, server does not see them.
In the chat log gui it would be easier to have the clients post their
own username a long with their message to the gui log. This is also
basic UDP, no reliability implemented yet. This will work by having the
driver class start a udp client and server when someone wants to send a
message rather than video chat.